requirements.txt
numpy>=1.21.0
opencv-python>=4.5.3
speechrecognition>=3.8.1
pyttsx3>=2.90
sounddevice>=0.4.4
scipy>=1.7.0
pillow>=8.3.1
pyaudio>=0.2.11
transformers>=4.30.0
torch>=2.0.0
pytesseract>=0.3.10
flask>=2.0.0
pymongo>=4.0.0
python-dotenv>=0.19.0
tensorflow>=2.10.0
tensorflow-hub>=0.12.0
tensorflow-text>=2.10.0 